+++ Tests/Bower/BowerTest.php
@@ -121,27 +121,18 @@ class BowerTest extends \PHPUnit_Framework_TestCase
         $this->bower->createDependencyMappingCache($config);
     }
 
+    /**
+     * @expectedException \Sp\BowerBundle\Bower\Exception\MappingException
+     */
     public function testCreateDependencyMappingCacheWithInvalidMapping()
     {
         $configDir = "/config_dir";
         $config = new Configuration($configDir);
 
-        $this->processBuilder->expects($this->at(1))->method('add')->with($this->equalTo($this->bin));
-        $this->processBuilder->expects($this->at(2))->method('add')->with($this->equalTo('list'));
-        $this->processBuilder->expects($this->at(3))->method('add')->with($this->equalTo('--map'));
-        $this->processBuilder->expects($this->once())->method('setWorkingDirectory')->with($this->equalTo($configDir));
         $this->processBuilder->expects($this->once())->method('getProcess')->will($this->returnValue($this->process));
         $this->process->expects($this->once())->method('isSuccessful')->will($this->returnValue(true));
 
-        $this->bower->expects($this->once())->method('dumpBowerConfig');
-        $this->eventDispatcher->expects($this->at(0))->method('dispatch')->with($this->equalTo(BowerEvents::PRE_EXEC));
-        $this->eventDispatcher->expects($this->at(1))->method('dispatch')->with($this->equalTo(BowerEvents::POST_EXEC));
-
-        $this->process->expects($this->once())->method('run')->with($this->equalTo(null));
-        $this->process->expects($this->once())->method('getOutput')->will($this->returnValue(""));
-
         $this->cache->expects($this->never())->method('save');
-        $this->cache->expects($this->once())->method('delete')->with($this->equalTo(hash('sha1', $configDir)));
 
         $this->bower->createDependencyMappingCache($config);
     }